rsu_identity_backend/requirements.txt
# 🇬🇦 RSU Gabon Backend - Dépendances Production (Windows Compatible)

# Django Core
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# Base de données
psycopg2-binary==2.9.7
redis==5.0.8
django-redis==5.4.0

# Configuration
python-decouple==3.8
dj-database-url==2.1.0

# Authentification & Sécurité
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
django-otp==1.2.4
django-ratelimit==4.1.0

# APIs & Documentation
drf-spectacular==0.26.5
requests==2.31.0

# Géolocalisation (sans PostGIS pour Windows)
geopy==2.4.0

# Machine Learning & IA
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Monitoring & Logging  
sentry-sdk==1.40.0
django-health-check==3.17.0
celery==5.3.4

# Utilitaires
pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7

# Développement
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

